requirements.txt
psycopg2-binary
Pillow
flask
flask_restful
flask_bcrypt
flask_cors
qrcode
PyJWT